packages/shared/pyproject.toml
[project]
name = "blackice-shared"
version = "0.1.0"
description = "Shared types and utilities for BLACKICE"
requires-python = ">=3.11"

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
